Change: Put Slider in its own folder again. Having trouble with it...
