+++ client/src/Components/CountryCard/Card.js
@@ -1,13 +1,9 @@
 import React from 'react';
-
 import { AppContextConsumer } from '../../AppContext';
-import CountryBorder from '../CountryBorder/CountryBorder';
-import Note from '../Note/Note';
-import FriendList from '../Friends/FriendList';
 import CardHeader from './CardHeader';
 import BorderBay from '../BorderBay/BorderBay';
-
-// import ReactModal from 'react-modal';
+import Note from '../Note/Note';
+import FriendList from '../Friends/FriendList';
 
 import './Card.css';
