+++ client/src/Components/Settings/Settings.less
@@ -12,10 +12,10 @@
   top: 100px;
   right: -300px;
   animation: Settings 0.5s forwards;
-  // animation-delay: 0.5s;
 
-}
-
-@keyframes Settings {
-  100% { right: 0; }
+  @keyframes Settings {
+    100% {
+      right: 0;
+    }
+  }
 }
